requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.47.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector-python==8.3.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
